// src/app/model/user.ts
export class userModel {
    userName: string;
    password: string;
    fullName: string;
    nickName: string;
    email: string;
    phone: string;
    birthDate: string;
    rePassword: string;
    rememberMe: string;
    terms: string;
}

export class userModelClass {
    userName: boolean;
    password: boolean;
    fullName: boolean;
    nickName: boolean;
    email: boolean;
    phone: boolean;
    birthDate: boolean;
    rePassword: boolean;
    rememberMe: boolean;
    terms: boolean;
}

export class userInformation {
    userName: boolean;
    fullName: boolean;
    nickName: boolean;
    email: boolean;
    phone: boolean;
    birthDate: boolean;
}